+++ server/index.js
@@ -1,5 +1,8 @@
 import express from 'express';
 import dotenv from 'dotenv';
+import request from 'request';
+import cors from 'cors'; 
+
 
 const router = express.Router(); 
 
@@ -11,19 +14,24 @@ var spotify_client_id = process.env.SPOTIFY_CLIENT_ID
 var spotify_client_secret = process.env.SPOTIFY_CLIENT_SECRET
 
 const app = express();
+app.use(cors());
 app.use(router); 
 
+var access_token = "";
 
-app.get('/auth/login', (req, res) => {
-});
 
-app.get('/auth/callback', (req, res) => {
-});
+// app.get('/auth/login', (req, res) => {
+// });
+
+// app.get('/auth/callback', (req, res) => {
+// });
 
 app.listen(port, () => {
   console.log(`Listening at http://localhost:${port}`)
 })
 
+
+console.log("up and running");
 var generateRandomString = function (length) {
     var text = '';
     var possible = 'ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789';
@@ -34,7 +42,15 @@ var generateRandomString = function (length) {
     return text;
   };
 
+app.get('/test', (req, res) => {
+  console.log('test route hit');
+  res.send('Test route response');
+});
+
 router.get('/auth/login', (req, res) => {
+
+  console.log('/auth/login endpoint accessed via router');
+
 
 var scope = "streaming \
                 user-read-email \
@@ -46,7 +62,7 @@ var auth_query_parameters = new URLSearchParams({
     response_type: "code",
     client_id: spotify_client_id,
     scope: scope,
-    redirect_uri: "http://localhost:5173/auth/callback",
+    redirect_uri: "http://localhost:5000/auth/callback",
     state: state
 })
 
@@ -55,13 +71,16 @@ res.redirect('https://accounts.spotify.com/authorize/?' + auth_query_parameters.
 
 app.get('/auth/callback', (req, res) => {
 
+  console.log('/auth/callback endpoint accessed via app');
+
+
     var code = req.query.code;
   
     var authOptions = {
       url: 'https://accounts.spotify.com/api/token',
       form: {
         code: code,
-        redirect_uri: "http://localhost:5173/auth/callback",
+        redirect_uri: "http://localhost:5000/auth/callback",
         grant_type: 'authorization_code'
       },
       headers: {
@@ -73,13 +92,15 @@ app.get('/auth/callback', (req, res) => {
   
     request.post(authOptions, function(error, response, body) {
       if (!error && response.statusCode === 200) {
-        var access_token = body.access_token;
-        res.redirect('/')
+        access_token = body.access_token;
+        res.redirect('http://localhost:5173/')
       }
     });
-  })
+  });
 
   app.get('/auth/token', (req, res) => {
+    console.log('/auth/token endpoint accessed');
+
     res.json(
        {
           access_token: access_token
