+++ src/App.jsx
@@ -10,7 +10,8 @@ function App() {
   useEffect(() => {
 
     async function getToken() {
-      const response = await fetch('/auth/token');
+      const response = await fetch('http://localhost:5000/auth/token');
+      console.log("Response:", response);
       const json = await response.json();
       setToken(json.access_token);
     }
@@ -19,39 +20,39 @@ function App() {
 
   }, []);
 
-  const [player, setPlayer] = useState(undefined);
+//   const [player, setPlayer] = useState(undefined);
 
-  useEffect(() => {
+//   useEffect(() => {
 
-    const script = document.createElement("script");
-    script.src = "https://sdk.scdn.co/spotify-player.js";
-    script.async = true;
+//     const script = document.createElement("script");
+//     script.src = "https://sdk.scdn.co/spotify-player.js";
+//     script.async = true;
 
-    document.body.appendChild(script);
+//     document.body.appendChild(script);
 
-    window.onSpotifyWebPlaybackSDKReady = () => {
+//     window.onSpotifyWebPlaybackSDKReady = () => {
 
-        const player = new window.Spotify.Player({
-            name: 'Web Playback SDK',
-            getOAuthToken: cb => { cb(props.token); },
-            volume: 0.5
-        });
+//         const player = new window.Spotify.Player({
+//             name: 'Web Playback SDK',
+//             getOAuthToken: cb => { cb(props.token); },
+//             volume: 0.5
+//         });
 
-        setPlayer(player);
+//         setPlayer(player);
 
-        player.addListener('ready', ({ device_id }) => {
-            console.log('Ready with Device ID', device_id);
-        });
+//         player.addListener('ready', ({ device_id }) => {
+//             console.log('Ready with Device ID', device_id);
+//         });
 
-        player.addListener('not_ready', ({ device_id }) => {
-            console.log('Device ID has gone offline', device_id);
-        });
+//         player.addListener('not_ready', ({ device_id }) => {
+//             console.log('Device ID has gone offline', device_id);
+//         });
 
 
-        player.connect();
+//         player.connect();
 
-    };
-}, []);
+//     };
+// }, []);
 
   return (
     <>
