+++ src/components/login.jsx
@@ -4,7 +4,7 @@ function Login() {
     return (
         <div className="App">
             <header className="App-header">
-                <a className="btn-spotify" href="/auth/login" >
+                <a className="btn-spotify" href="http://localhost:5000/auth/login" >
                     Login with Spotify 
                 </a>
             </header>
